test(agent): stabilize historical Side validation
