+++ packages/agent/host/types.go
@@ -1070,17 +1070,3 @@ type FenceGoalGenerationResult struct {
 	IntentAccepted bool
 	Settled        bool
 }
-
-type GoalReconcileRequiredInput struct {
-	WorkspaceID         string
-	AgentSessionID      string
-	RequestID           string
-	ProviderTurnID      string
-	Reason              string
-	FenceMode           string
-	ExpectedOperationID string
-	ExpectedRevision    int64
-	ExpectedRepairEpoch int64
-	QuiesceSucceeded    bool
-	QuiesceError        string
-}
